Hoist deletePhotoByUri
